애플리케이션 대신 DB로 기능 구현

+++ Spring/Ko-pring/library-app/src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -2,4 +2,4 @@ package com.group.libraryapp.dto.book.response
 
 import com.group.libraryapp.domain.book.BookType
 
-data class BookStatResponse(val type: BookType, val count: Int)
+data class BookStatResponse(val type: BookType, val count: Long)
